requirements.txt
line-bot-sdk
flask
gunicorn
gensim
jieba
numpy
beautifulsoup4
requests
psycopg2
Flask-SQLAlchemy
Flask-Migrate
click
googlemaps
